+++ apps/backend/src/app/.well-known/jwks.json/route.ts
@@ -1,4 +1,5 @@
 import { yupArray, yupNumber, yupObject, yupString } from "@stackframe/stack-shared/dist/schema-fields";
+import { getEnvVariable } from "@stackframe/stack-shared/dist/utils/env";
 import { getPublicJwkSet } from "@stackframe/stack-shared/dist/utils/jwt";
 import { createSmartRouteHandler } from "../../../route-handlers/smart-route-handler";
 
@@ -15,13 +16,17 @@ export const GET = createSmartRouteHandler({
     bodyType: yupString().oneOf(["json"]).required(),
     body: yupObject({
       keys: yupArray().required(),
+      message: yupString().optional(),
     }).required(),
   }),
   async handler() {
     return {
       statusCode: 200,
       bodyType: "json",
-      body: await getPublicJwkSet(),
+      body: {
+        ...await getPublicJwkSet(getEnvVariable("STACK_SERVER_SECRET")),
+        message: "This is deprecated. Please disable the legacy JWT signing in the project setting page, and move to /api/v1/projects/<project-id>/.well-known/jwks.json",
+      }
     };
   },
 });