+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @stackframe/stack-dashboard
 
+## 2.6.3
+
+### Patch Changes
+
+- Bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.3
+  - @stackframe/stack@2.6.3
+  - @stackframe/stack-emails@2.6.3
+  - @stackframe/stack-ui@2.6.3
+
 ## 2.6.2
 
 ### Patch Changes