+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,13 @@
 # e2e-tests
 
+## 2.6.3
+
+### Patch Changes
+
+- Bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.3
+
 ## 2.6.2
 
 ### Patch Changes