+++ examples/demo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # demo-app
 
+## 2.6.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.3
+  - @stackframe/stack@2.6.3
+  - @stackframe/stack-ui@2.6.3
+
 ## 2.6.2
 
 ### Patch Changes